Added Carts and Products Event Integrations

+++ internal/services/cart_service/internal/carts/features/updating_cart/v1/update_cart_endpoint.go
@@ -2,13 +2,18 @@ package v1
 
 import (
 	"context"
+	"encoding/json"
 	"net/http"
 
 	v "github.com/RussellLuo/validating/v3"
+	"github.com/ThreeDotsLabs/watermill"
+	"github.com/ThreeDotsLabs/watermill/message"
 	"github.com/danielgtaylor/huma/v2"
+	"github.com/tguankheng016/go-ecommerce-microservice/internal/pkg/events"
 	httpServer "github.com/tguankheng016/go-ecommerce-microservice/internal/pkg/http"
 	"github.com/tguankheng016/go-ecommerce-microservice/internal/pkg/permissions"
 	cartModel "github.com/tguankheng016/go-ecommerce-microservice/internal/services/cart_service/internal/carts/models"
+	productModel "github.com/tguankheng016/go-ecommerce-microservice/internal/services/cart_service/internal/products/models"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 )
@@ -37,6 +42,7 @@ func (e HumaUpdateCartRequest) Schema() v.Schema {
 func MapRoute(
 	api huma.API,
 	database *mongo.Database,
+	publisher message.Publisher,
 ) {
 	huma.Register(
 		api,
@@ -54,11 +60,11 @@ func MapRoute(
 				permissions.Authorize(api, ""),
 			},
 		},
-		updateCart(database),
+		updateCart(database, publisher),
 	)
 }
 
-func updateCart(database *mongo.Database) func(context.Context, *HumaUpdateCartRequest) (*struct{}, error) {
+func updateCart(database *mongo.Database, publisher message.Publisher) func(context.Context, *HumaUpdateCartRequest) (*struct{}, error) {
 	return func(ctx context.Context, request *HumaUpdateCartRequest) (*struct{}, error) {
 		errs := v.Validate(request.Schema())
 		for _, err := range errs {
@@ -70,14 +76,18 @@ func updateCart(database *mongo.Database) func(context.Context, *HumaUpdateCartR
 			return nil, huma.Error401Unauthorized("unable to get current user id")
 		}
 
+		productCollection := database.Collection(productModel.ProductCollectionName)
 		cartCollection := database.Collection(cartModel.CartCollectionName)
 
 		filter := bson.D{
 			bson.E{Key: "id", Value: request.Body.Id},
-			bson.E{Key: "userid", Value: userId},
+			bson.E{Key: "user_id", Value: userId},
 		}
 
-		err := cartCollection.FindOne(ctx, filter).Err()
+		quantityChanged := 0
+
+		var cart cartModel.Cart
+		err := cartCollection.FindOne(ctx, filter).Decode(&cart)
 
 		if err != nil && err != mongo.ErrNoDocuments {
 			return nil, huma.Error500InternalServerError(err.Error())
@@ -87,12 +97,39 @@ func updateCart(database *mongo.Database) func(context.Context, *HumaUpdateCartR
 			return nil, huma.Error400BadRequest("invalid cart id")
 		}
 
+		productFilter := bson.D{bson.E{Key: "id", Value: cart.ProductId}}
+
+		var product productModel.Product
+		err = productCollection.FindOne(ctx, productFilter).Decode(&product)
+
+		if err != nil && err != mongo.ErrNoDocuments {
+			return nil, huma.Error500InternalServerError(err.Error())
+		}
+
+		if err == mongo.ErrNoDocuments {
+			return nil, huma.Error400BadRequest("invalid product id")
+		}
+
 		if request.Body.Quantity == 0 {
+			quantityChanged = cart.Quantity
+
 			_, err := cartCollection.DeleteOne(ctx, filter)
 			if err != nil {
 				return nil, huma.Error500InternalServerError(err.Error())
 			}
-		} else {
+		} else if cart.Quantity != request.Body.Quantity {
+			quantityChanged = cart.Quantity - request.Body.Quantity
+
+			if quantityChanged < 0 {
+				if product.StockQuantity == 0 {
+					return nil, huma.Error400BadRequest("this product is out of stock.")
+				}
+
+				if (-1 * quantityChanged) > product.StockQuantity {
+					return nil, huma.Error400BadRequest("the selected quantity exceeds the remaining stock.")
+				}
+			}
+
 			update := bson.D{
 				{
 					Key: "$set",
@@ -106,6 +143,21 @@ func updateCart(database *mongo.Database) func(context.Context, *HumaUpdateCartR
 			if err != nil {
 				return nil, huma.Error500InternalServerError(err.Error())
 			}
+		}
+
+		if quantityChanged != 0 {
+			changeProductQuantityEvent := events.ChangeProductQuantityEvent{
+				Id:              cart.ProductId,
+				QuantityChanged: quantityChanged,
+			}
+
+			payload, err := json.Marshal(changeProductQuantityEvent)
+			if err != nil {
+				return nil, huma.Error500InternalServerError(err.Error())
+			}
+
+			msg := message.NewMessage(watermill.NewUUID(), payload)
+			publisher.Publish(events.ChangeProductQuantityTopicV1, msg)
 		}
 
 		return nil, nil
